Movied three methods form UserContorller to IdentityController
